# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/organization/trips": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "get all trips",
                "description": "get all trips",
                "operationId": "get all trips",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Trips"
                ],
                "summary": "create a trip",
                "description": "create a trip",
                "operationId": "create a trip",
                "parameters": [
                    {
                        "name": "path_id",
                        "in": "path",
                        "description": "path id",
                        "required": true
                    },
                    {
                        "name": "repitition",
                        "in": "path",
                        "description": "repitition",
                        "required": true
                    },
                    {
                        "name": "date",
                        "in": "path",
                        "description": "date",
                        "required": true
                    },
                    {
                        "name": "time",
                        "in": "path",
                        "description": "time",
                        "required": true
                    },
                    {
                        "name": "num_seats",
                        "in": "path",
                        "description": "number of seats",
                        "required": true
                    },
                    {
                        "name": "public",
                        "in": "path",
                        "description": "is the trip public 0 or 1",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/organization/trips/pay/{id}": {
            "post": {
                "tags": [
                    "Trips"
                ],
                "summary": "pay a trip",
                "description": "pay a trip",
                "operationId": "pay a trip",
                "parameters": [
                    {
                        "name": "payment_method",
                        "in": "path",
                        "description": "payment method (wallet or credit)",
                        "required": true
                    },
                    {
                        "name": "card_number",
                        "in": "path",
                        "description": "card number(for credit payment method)",
                        "required": true
                    },
                    {
                        "name": "exp_month",
                        "in": "path",
                        "description": "expiration month (for credit payment method)",
                        "required": true
                    },
                    {
                        "name": "exp_year",
                        "in": "path",
                        "description": "expiration year (for credit payment method)",
                        "required": true
                    },
                    {
                        "name": "CVC",
                        "in": "path",
                        "description": "CVC (three-digit number) (for credit payment method)",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/organization/trips/users/{id}": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "get users in a trip",
                "description": "get users in a trip",
                "operationId": "get users in a trip",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/organization/trips/{id}": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "get trip by id",
                "description": "get trip by id",
                "operationId": "get trip by id",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Trips"
                ],
                "summary": "delete trip by id",
                "description": "delete trip by id",
                "operationId": "delete trip by id",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}